requirements.txt
PyYAML>=6.0
requests>=2.28.0
GitPython>=3.1.30
